requirements_windows_fix.txt
# Streamlit App Builder - Windows Compatible Requirements
# This avoids the Pillow build issue on Windows

# Core dependencies
streamlit>=1.36.0
google-generativeai>=0.3.0  # For Gemini API
python-dotenv>=1.0.0

# Image handling - using compatible versions
Pillow>=10.0.0,<11.0.0  # Flexible version that has pre-built wheels

# Optional but recommended
streamlit-image-select>=0.6.0
streamlit-extras>=0.3.6

# For enhanced UI (optional)
streamlit-option-menu>=0.3.6
streamlit-antd-components>=0.3.0

# Monitoring (optional)
wandb>=0.16.0  # For Weights & Biases tracking